fix mentor's comments
