+++ debugscripts/post_json.restbook
@@ -1,0 +1,1 @@
+[{"kind":2,"language":"rest-book","value":"POST localhost:44751/update\nContent-Type: application/json \n\n{\"id\":\"GetSet312\",\"type\":\"counter\",\"delta\":123336}","outputs":[{"mime":"x-application/rest-book","value":{"status":200,"statusText":"OK","headers":{"Date":"Thu, 02 Jun 2022 19:04:31 GMT","Content-Type":"text/html; charset=utf-8","Content-Length":"32","Connection":"close"},"config":{"timeout":10000,"xsrfCookieName":"XSRF-TOKEN","xsrfHeaderName":"X-XSRF-TOKEN","headers":{"Accept":"application/json, text/plain, */*","Content-Type":"application/json","User-Agent":"axios/0.21.1","Content-Length":50}},"request":{"method":"POST","httpVersion":"1.1","responseUrl":"http://localhost:44751/update","timeout":10000,"headers":{"Content-Type":"application/json"},"data":{"id":"GetSet312","type":"counter","delta":123336}},"data":"<h1>Counter metric</h1>GetSet312"}},{"mime":"text/x-json","value":{"status":200,"statusText":"OK","headers":{"Date":"Thu, 02 Jun 2022 19:04:31 GMT","Content-Type":"text/html; charset=utf-8","Content-Length":"32","Connection":"close"},"config":{"timeout":10000,"xsrfCookieName":"XSRF-TOKEN","xsrfHeaderName":"X-XSRF-TOKEN","headers":{"Accept":"application/json, text/plain, */*","Content-Type":"application/json","User-Agent":"axios/0.21.1","Content-Length":50}},"request":{"method":"POST","httpVersion":"1.1","responseUrl":"http://localhost:44751/update","timeout":10000,"headers":{"Content-Type":"application/json"},"data":{"id":"GetSet312","type":"counter","delta":123336}},"data":"<h1>Counter metric</h1>GetSet312"}},{"mime":"text/html","value":"<h1>Counter metric</h1>GetSet312"}]},{"kind":2,"language":"rest-book","value":"POST localhost:44751/value\nContent-Type: application/json \n\n{\"id\":\"GetSet312\",\"type\":\"counter\"}","outputs":[{"mime":"x-application/rest-book","value":{"status":200,"statusText":"OK","headers":{"Date":"Thu, 02 Jun 2022 19:04:34 GMT","Content-Type":"application/json","Content-Length":"50","Connection":"close"},"config":{"timeout":10000,"xsrfCookieName":"XSRF-TOKEN","xsrfHeaderName":"X-XSRF-TOKEN","headers":{"Accept":"application/json, text/plain, */*","Content-Type":"application/json","User-Agent":"axios/0.21.1","Content-Length":35}},"request":{"method":"POST","httpVersion":"1.1","responseUrl":"http://localhost:44751/value","timeout":10000,"headers":{"Content-Type":"application/json"},"data":{"id":"GetSet312","type":"counter"}},"data":{"id":"GetSet312","type":"counter","delta":123336}}},{"mime":"text/x-json","value":{"status":200,"statusText":"OK","headers":{"Date":"Thu, 02 Jun 2022 19:04:34 GMT","Content-Type":"application/json","Content-Length":"50","Connection":"close"},"config":{"timeout":10000,"xsrfCookieName":"XSRF-TOKEN","xsrfHeaderName":"X-XSRF-TOKEN","headers":{"Accept":"application/json, text/plain, */*","Content-Type":"application/json","User-Agent":"axios/0.21.1","Content-Length":35}},"request":{"method":"POST","httpVersion":"1.1","responseUrl":"http://localhost:44751/value","timeout":10000,"headers":{"Content-Type":"application/json"},"data":{"id":"GetSet312","type":"counter"}},"data":{"id":"GetSet312","type":"counter","delta":123336}}},{"mime":"text/html","value":"[object Object]"}]}]
