+++ server.cmd
