+++ server.sh
@@ -5,7 +5,7 @@ export POLL_INTERVAL=2s
 export RESTORE=false
 export SHUTDOWN_TIMEOUT=5s
 export STORE_INTERVAL=10s
-export STORE_FILE=.\devops-metrics-db.json
+export STORE_FILE=./devops-metrics-db.json
 export RESTORE=true
 #export KEY=/tmp/zK3deX0
 export DATABASE_DSN=postgres://postgres:qweasd@localhost:5432/yandex?sslmode=disable
